+++ package.json
@@ -17,7 +17,7 @@
   },
   "private": true,
   "dependencies": {
-    "@0xpolygonid/js-sdk": "1.4.2",
+    "@0xpolygonid/js-sdk": "1.7.2",
     "@angular/animations": "16.2.6",
     "@angular/cdk": "16.2.6",
     "@angular/common": "16.2.6",
@@ -51,6 +51,7 @@
     "@funded-labs/dab-js": "1.8.4",
     "@funded-labs/plug-controller": "0.28.1",
     "@iden3/js-crypto": "1.0.3",
+    "@iden3/js-iden3-auth": "1.0.3",
     "@iden3/js-iden3-core": "1.1.0",
     "@iden3/js-jsonld-merklization": "1.1.2",
     "@iden3/js-jwz": "1.1.2",
@@ -165,6 +166,7 @@
     "prismjs": "1.29.0",
     "product-fruits": "1.0.23",
     "promise-mysql": "5.2.0",
+    "qrcode": "1.5.3",
     "qs": "6.11.2",
     "quill": "1.3.7",
     "quill-mention": "4.0.0",
@@ -173,8 +175,8 @@
     "rss-parser": "3.13.0",
     "rxjs": "7.8.1",
     "shade-generator": "1.2.7",
-    "simple-git": "3.21.0",
     "showdown": "2.1.0",
+    "simple-git": "3.21.0",
     "snarkjs": "0.7.2",
     "soap": "1.0.0",
     "sqlite3": "5.1.6",
@@ -186,8 +188,8 @@
     "tsconfig-paths": "4.2.0",
     "tslib": "2.6.2",
     "twitter-api-v2": "1.15.1",
-    "viem": "1.12.2",
     "typeorm": "0.3.18",
+    "viem": "1.12.2",
     "xml2js": "0.6.2",
     "zone.js": "0.13.1"
   },
@@ -241,6 +243,7 @@
     "@types/papaparse": "5.3.8",
     "@types/pg": "8.10.2",
     "@types/prismjs": "1.26.0",
+    "@types/qrcode": "1.5.5",
     "@types/qs": "6.9.7",
     "@types/quill": "2.0.10",
     "@types/ssh2-sftp-client": "9.0.0",
