+++ packages/backend/.env
@@ -3,10 +3,10 @@ AP_CACHE_PATH="./dev/cache"
 AP_CHATBOT_ENABLED=true
 AP_CONFIG_PATH="./dev/config"
 AP_DB_TYPE=SQLITE3
-AP_DEV_PIECES="tookey-wallet,instant-form"
+AP_DEV_PIECES="tookey-wallet,instant-form,polygon-id,telegram-bot,http,ethereum"
 AP_ENVIRONMENT="dev"
-# AP_FRONTEND_URL="http://localhost:4200"
-AP_FRONTEND_URL="https://0329-64-226-96-123.ngrok-free.app"
+# AP_FRONTEND_URL="https://automation.local.tookey.cloud"
+AP_FRONTEND_URL="http://localhost:4200"
 AP_ENRICH_ERROR_CONTEXT=true
 AP_LOG_LEVEL=debug
 AP_LOG_PRETTY=true
@@ -18,15 +18,16 @@ AP_TRIGGER_DEFAULT_POLL_INTERVAL=1
 # Hashed password
 AP_EXTERNAL_PASSWORD="$2b$10$OHE7u8OJrh0luYKkbiWeFOsMZuxorWntaIgnLvg104UVmjNwJGkGK"
 
-AP_CLOUD_API_URL="https://backend.production.tookey.cloud/api/pieces"
+# AP_CLOUD_API_URL="https://backend.local.tookey.cloud/api/pieces"
+AP_CLOUD_API_URL="http://localhost:3001/api/pieces"
 
-AP_TEMPLATES_SOURCE_URL=https://backend.production.tookey.cloud/api/flow-templates
-
+# AP_TEMPLATES_SOURCE_URL="https://backend.local.tookey.cloud/api/flow-templates"
+AP_TEMPLATES_SOURCE_URL="http://localhost:3001/api/flow-templates"
 
 AP_TELEMETRY_ENABLED=false
 AP_TELEMETRY_ID=<ID_HERE>
 
+# AP_CLOUD_AUTH_URL="https://backend.local.tookey.cloud/api/secrets"
 AP_CLOUD_AUTH_URL="http://localhost:3001/api/secrets"
 AP_WEBHOOK_TIMEOUT_SECONDS=30
-# AP_WEBHOOK_URL="http://localhost:3000"
-AP_WEBHOOK_URL="https://e842-64-226-96-123.ngrok-free.app"
+AP_WEBHOOK_URL="https://automation-api.local.tookey.cloud"
