+++ packages/backend/project.json
@@ -91,18 +91,6 @@
     },
     "test:docker": {
       "command": "UID=\"$(id -u)\" GID=\"$(id -g)\" docker compose --profile full -f docker-compose.test.yml up --exit-code-from app --attach app"
-    },
-    "db": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "ts-node -r tsconfig-paths/register -P packages/backend/tsconfig.app.json ./node_modules/typeorm/cli.js"
-      }
-    },
-    "db-migration": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "nx db backend -- migration:generate -p -d packages/backend/src/app/database/database-connection.ts packages/backend/src/app/database/migration/{args.name}"
-      }
     }
   },
   "tags": []
