+++ packages/pieces/polygon-id/package.json
@@ -1,4 +1,4 @@
 {
     "name": "@tookey-io/piece-polygon-id",
-    "version": "0.0.1"
+    "version": "0.2.1"
 }
