+++ packages/pieces/polygon-id/src/lib/constants.ts
@@ -1,0 +1,4 @@
+
+export const STORAGE_PREFIX = 'polygon_id_3';
+export const STORAGE_IDEN3_URL = STORAGE_PREFIX + '_iden3_callback_url';
+export const DELIMITER = '__!__';
