+++ packages/pieces/polygon-id/src/lib/defined.ts
@@ -1,0 +1,2 @@
+
+export const defined = <T>(x: T | undefined): x is T => x !== undefined;
