+++ packages/pieces/telegram-bot/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "@activepieces/piece-telegram-bot",
-  "version": "0.3.8"
+  "name": "@tookey-io/piece-telegram-bot",
+  "version": "1.0.0"
 }
