+++ packages/pieces/telegram-bot/project.json
@@ -25,5 +25,7 @@
       }
     }
   },
-  "tags": []
+  "tags": [
+    "@tookey-io"
+  ]
 }
